Moved helpers from toasting.ts
