+++ src/pages/_app.tsx
@@ -10,8 +10,9 @@ import { loadAllOptionsWithAsset } from "../datasources/options";
 import { NFTOpt } from "../../typechain-types";
 import NFTOptSolContract from "../../artifacts/contracts/NFTOpt.sol/NFTOpt.json";
 import addresses from "../../addresses.json";
-import { dismissLastToast, actionLabels, TOAST_DURATION, events } from "../utils/toasting";
+import { dismissLastToast, TOAST_DURATION } from "../utils/toasting";
 import { BigNumber } from "ethers";
+import { actionLabels, events } from "../utils/frontend/labels";
 
 const OptionsHashContext       = createContext(0);
 const AccountContext           = createContext("");
